Addressed code review.
