+++ python/ray/dashboard/metrics_exporter/client.py
@@ -1,8 +1,9 @@
 import logging
+import threading
 import traceback
+import time
 
 from ray.dashboard.metrics_exporter import api
-from ray.dashboard.metrics_exporter.exporter import Exporter
 
 logger = logging.getLogger(__name__)
 
@@ -14,14 +15,14 @@ class MetricsExportClient:
     multiple threads that export the same metrics.
 
     Args:
-        address: Address to export metrics
+        address(str): Address to export metrics
         dashboard_controller(BaseDashboardController): Dashboard controller to
             run dashboard business logic.
         dashboard_id(str): Unique dashboard ID.
+        exporter(Exporter): Thread to export metrics.
     """
 
-    def __init__(self, address, dashboard_controller, dashboard_id,
-                 exporter: Exporter):
+    def __init__(self, address, dashboard_controller, dashboard_id, exporter):
         self.dashboard_id = dashboard_id
         self.auth_url = "{}/auth".format(address)
         self.dashboard_controller = dashboard_controller
@@ -75,9 +76,65 @@ class MetricsExportClient:
                 logger.error(error)
                 return False, error
 
-        # Exporter is a Python thread that keeps exporting metrics with
-        # access token obtained by an authentication process.
         self.exporter.access_token = self.auth_info.access_token
         self.exporter.start()
         self.is_exporting_started = True
         return True, None
+
+
+class Exporter(threading.Thread):
+    """Python thread that exports metrics periodically.
+
+    Args:
+        dashboard_id(str): Unique Dashboard ID.
+        address(str): Address to export metrics.
+        dashboard_controller(BaseDashboardController): dashboard
+            controller for dashboard business logic.
+        update_frequency(float): Frequency to export metrics.
+    """
+
+    def __init__(self,
+                 dashboard_id,
+                 address,
+                 dashboard_controller,
+                 update_frequency=1.0):
+        assert update_frequency >= 1.0
+
+        self.dashboard_id = dashboard_id
+        self.dashboard_controller = dashboard_controller
+        self.export_address = "{}/ingest".format(address)
+        self.update_frequency = update_frequency
+        self._access_token = None
+
+        super().__init__()
+
+    @property
+    def access_token(self):
+        return self._access_token
+
+    @access_token.setter
+    def access_token(self, access_token):
+        self._access_token = access_token
+
+    def export(self, ray_config, node_info, raylet_info, tune_info,
+               tune_availability):
+        api.ingest_request(self.export_address, self.dashboard_id,
+                           self.access_token, ray_config, node_info,
+                           raylet_info, tune_info, tune_availability)
+        # TODO(sang): Add piggybacking response handler.
+
+    def run(self):
+        assert self.access_token is not None, (
+            "Set access token before running an exporter thread.")
+        while True:
+            try:
+                time.sleep(self.update_frequency)
+                self.export(self.dashboard_controller.get_ray_config(),
+                            self.dashboard_controller.get_node_info(),
+                            self.dashboard_controller.get_raylet_info(),
+                            self.dashboard_controller.tune_info(),
+                            self.dashboard_controller.tune_availability())
+            except Exception as e:
+                logger.error("Exception occured while exporting metrics: {}.\n"
+                             "Traceback: {}".format(e, traceback.format_exc()))
+                continue
